+++ orm/src/main/java/org/hibernate/tool/internal/export/ddl/Hbm2DDLExporter.java
@@ -14,7 +14,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.hibernate.tool.hbm2x;
+package org.hibernate.tool.internal.export.ddl;
 
 import java.io.File;
 import java.util.EnumSet;
